Adding missing corrections

+++ olaf/pipeline/pipeline_component/candidate_term_enrichment/semantic_based_enrichment.py
@@ -26,7 +26,7 @@ class SemanticBasedEnrichment(PipelineComponent):
 
         Parameters
         ----------
-        threshold : int, optional
+        threshold : float, optional
             The threshold defines the minimum similarity score required to be synonymous.
             By default the threshold is set to 0.9.
         """
